Refactor go code into handlers and backend
